fixed some user creation stuff
Put password setting on a seperate form.  There is no way (that I have
found) to check a password for vailidity without just attempting to set
it.  I put password setting in it's own form on it's own page as a
result.

The user creation page does not switch to password setting
automatically.  This is a bug.